+++ mangadap/datacube/manga.py
@@ -27,7 +27,6 @@ from ..util.drpfits import DRPFits, DRPFitsBitMask
 from ..util.parser import DefaultConfig
 from ..util.constants import DAPConstants
 from ..util.covariance import Covariance
-from ..util.filter import interpolate_masked_vector
 from ..spectra import MaNGARSS
 from .datacube import DataCube
 
@@ -265,7 +264,7 @@ class MaNGADataCube(DRPFits, DataCube):
         """
         path, file_name = cls.default_paths(int(plate), int(ifudesign), log=log, drpver=drpver,
                                             redux_path=redux_path, directory_path=directory_path)
-        return cls(os.path.join(directory_path, file_name), **kwargs)
+        return cls(os.path.join(path, file_name), **kwargs)
 
     def load_rss(self, force=False):
         """