+++ mangadap/proc/emissionlinemodel.py
@@ -39,6 +39,8 @@ import glob
 import logging
 import warnings
 
+from IPython import embed
+
 import numpy
 
 from scipy import interpolate