+++ mangadap/proc/sasuke.py
@@ -41,6 +41,8 @@ import warnings
 import logging
 import warnings
 
+from IPython import embed
+
 import numpy
 from scipy import interpolate, fftpack