+++ mangadap/tests/test_drpcomplete.py
@@ -16,6 +16,8 @@ def test_write_cfg():
         os.remove(ofile)
 
     # Read the test DRPComplete file.
+    # TODO: Temporarily override
+    drp_test_version = 'v2_7_1'
     drpc = DRPComplete(drpver=drp_test_version, directory_path=remote_data_file(), readonly=True)
 
     # Write the base-level configuration file