+++ mangadap/util/drpfits.py
@@ -94,6 +94,7 @@ Revision history
 """
 import time
 import os
+import warnings
 
 from configparser import ConfigParser
 
@@ -201,34 +202,38 @@ class DRPFits:
         """
         Determine the spectral resolution channel to use.
 
-        Precedence follows this order: ``PREDISP``, ``PRESPECRES``,
-        ``DISP``, ``SPECRES``.
+        Precedence follows this order: ``LSFPRE``, ``PRESPECRES``,
+        ``LSFPOST``, ``SPECRES``.
 
         Args:
             hdu (`astropy.io.fits.HDUList`):
                 The opened MaNGA DRP file.
             ext (:obj:`str`, optional):
                 Specify the extension with the spectral estimate to
-                use. Should be in None, ``PREDISP``, ``PRESPECRES``,
-                ``DISP``, or ``SPECRES``. The default is None, which
-                means it will return the extension found first in the
-                order above. None is returned if none of the
+                use. Should be in None, ``LSFPRE``, ``PRESPECRES``,
+                ``LSFPOST``, or ``SPECRES``. The default is None,
+                which means it will return the extension found first
+                in the order above. None is returned if none of the
                 extensions are present.
 
         Returns:
             :obj:`str`: The name of the preferred extension to use.
         """
-        available = [h.name for h in hdu if h.name in ['PREDISP', 'DISP', 'PRESPECRES', 'SPECRES']]
+        allowed = ['LSFPRE', 'LSFPOST', 'PRESPECRES', 'SPECRES']
+        if ext is not None and ext not in allowed:
+            warnings.warn('{0} is not a viable spectral resolution extension.'.format(ext))
+            return None
+        available = [h.name for h in hdu if h.name in allowed]
         _ext = ext
         if ext is None:
-            _ext = 'PREDISP'
+            _ext = 'LSFPRE'
             if _ext not in available:
                 _ext = 'PRESPECRES'
             if _ext not in available:
-                _ext = 'DISP'
+                _ext = 'LSFPOST'
             if _ext not in available:
                 _ext = 'SPECRES'
-        return None if _ext not in available else _ext
+        return _ext
 
     @staticmethod
     def spectral_resolution(hdu, ext=None, fill=False, median=False):
@@ -249,12 +254,12 @@ class DRPFits:
                 leave masked pixels in returned array.
             median (:obj:`bool`, optional):
                 Return a single vector with the median spectral
-                resolution instead of a per spectrum array. When
-                using the ``SPECRES`` extension, this just returns
-                the vector provided by the DRP file; when using the
-                ``DISP`` extension, this performs a masked median
-                across the array and then interpolates any
-                wavelengths that were masked in all vectors.
+                resolution instead of a per spectrum array. When using
+                the ``SPECRES`` extension, this just returns the vector
+                provided by the DRP file; when using either of the
+                ``LSF`` extensions, this performs a masked median across
+                the array and then interpolates any wavelengths that
+                were masked in all vectors.
 
         Returns:
             :obj:`tuple`: Returns a :obj:`str` with the name of the
@@ -272,19 +277,20 @@ class DRPFits:
             raise ValueError('No valid spectral resolution extension.')
         if ext is not None and _ext is None:
             raise ValueError('No extension: {0}'.format(ext))
+        warnings.warn('Extension {0} used to define spectral resolution.'.format(_ext))
             
+        # Set the mode based on the shape of the flux extension
+        mode = 'CUBE' if hdu['FLUX'].data.ndim == 3 else 'RSS'
+
         # Build the spectral resolution vectors
         if 'SPECRES' in _ext:
             sres = numpy.ma.MaskedArray(hdu[_ext].data, mask=numpy.invert(hdu[_ext].data > 0))
             if fill:
                 sres = numpy.ma.MaskedArray(interpolate_masked_vector(sres))
             if not median:
-                nspec = hdu['FLUX'].data.shape[0]
-                sres = numpy.ma.tile(sres, (nspec,1))
+                sres = numpy.tile(sres, (hdu['FLUX'].data.shape[0],1)) if mode == 'RSS' \
+                         else numpy.tile(sres, (*hdu['FLUX'].data.shape[1:][::-1],1)).T
             return _ext, sres
-
-        # Set the mode based on the shape of the flux extension
-        mode = 'CUBE' if hdu['FLUX'].data.ndim == 3 else 'RSS'
 
         # Otherwise dealing with the DISP data
         sres = numpy.ma.MaskedArray(hdu[_ext].data)